requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.1
cycler==0.12.1
efficientnet-pytorch==0.7.1
filelock==3.9.0
fonttools==4.43.1
fsspec==2023.4.0
grad-cam==1.4.8
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.28.1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tqdm==4.66.1
triton==2.1.0
ttach==0.0.3
typing_extensions==4.4.0
urllib3==1.26.13
